Fix building with clang
Fix error: non-constant-expression cannot be narrowed from type 'size_t'
(aka 'unsigned long') to 'int' in initializer list [-Wc++11-narrowing]

+++ src/app.cpp
@@ -166,7 +166,7 @@ int16_t input_state(unsigned port, unsigned device, unsigned index, unsigned id)
     return 0;
 }
 
-void App::videoRefresh(const void *data, unsigned width, unsigned height, size_t pitch) {
+void App::videoRefresh(const void *data, unsigned width, unsigned height, int pitch) {
     QImage i = QImage{reinterpret_cast<const uchar*>(data), static_cast<int>(width), static_cast<int>(height), pitch, m_imageFormat};
     if(m_imageFormat == QImage::Format_RGBX8888)
         i = i.rgbSwapped();
